fix stow

+++ stow.sh
@@ -1,2 +1,7 @@
-rm ~/.zshrc
+#!/bin/bash
+
+if [ -e "~/.zshrc" ]; then
+	rm ~/.zshrc
+fi
+
 stow . -t ~/
